LineageRegistrationAlgorithmTest: extract tests for LineageRegistrationUtils
